backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-docx
lxml
langchain
langchain-openai
langchain-anthropic
arq
redis
chromadb
jinja2
python-multipart
